+++ deploy/roles/filter/templates/filter.service
@@ -6,9 +6,9 @@ After=network.target
 Type=simple
 User={{ ansible_user_id }}
 WorkingDirectory={{ ansible_env.HOME }}/{{ git_name}}/pipeline/filter
-ExecStartPre=/usr/bin/touch {{ ansible_env.HOME }}/lockfile
+#ExecStartPre=/usr/bin/touch {{ ansible_env.HOME }}/lockfile
 ExecStart=/usr/bin/python3 filter_runner.py
-ExecStopPost=/usr/bin/rm /home/ubuntu/lockfile
+#ExecStopPost=/usr/bin/rm /home/ubuntu/lockfile
 Restart=on-failure
